# CODE/docs/referencias/liwa_openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LIWA Empresas API",
    "description": "LIWA Empresas API SMS permite a nuestros clientes enviar de forma segura y eficiente mensajes de texto. Hemos habilitado destinos internacionales en nuestra nueva versión. Se requiere la utilización de JWT para la autenticación y uso de un API Key mediante el backoffice. Puedes agregar la IP 0.0.0.0 para consumirla desde un servidor sin ip fija.",
    "termsOfService": "https://www.cellvoz.com/wp-content/uploads/2021/03/Terminos-y-Condiciones-022021.pdf",
    "contact": {
      "email": "contacto@cellvoz.com.co"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://api.liwa.co/v2",
      "description": "Production"
    },
    {
      "url": "https://api-dev.liwa.co/v2",
      "description": "Staging"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Obtener Token de autorización",
        "operationId": "login",
        "requestBody": {
          "description": "Cuenta y clave del cliente LIWA",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Credenciales validas, se obtiene el token que debe ser enviado en cada transacción segura. El token de autorización tiene una duración de 24 horas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Credenciales invalidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseApi"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Login": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string",
            "description": "Numero de cuenta Liwa"
          },
          "password": {
            "type": "string",
            "description": "Contraseña de cuenta Liwa"
          }
        }
      },
      "ErrorResponseApi": {
        "type": "object",
        "properties": {
          "succes": {
            "type": "boolean",
            "default": false
          },
          "message": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "api_key": {
        "type": "apiKey",
        "name": "api-key",
        "in": "header"
      }
    }
  }
}